Add the `root-trees.sh` module to support rt management
Add a new module (`src/lib/root-trees.sh` to support the management of
root trees. In the future the VM concept will be replaced by the
root-tree concept either completely or by leaving for the VM only
hardware properties (to be decided).

A root tree is a tree used as root when booting a certain kernel.

Signed-off-by: Ágatha Isabelle Chris Moreira Guedes <code@agatha.dev>